[deps] Add 'rollup-plugin-sourcemaps' intended for release builds
We plan to enable source maps for release builds behind a GN flag. For
this to work rollup needs to consume the source maps already generated
by tsc and merge it with its own source map. Rollup can't do this
natively so this CL adds a plugin that can.

R=jacktfranklin@chromium.org

Bug: 1430396
Change-Id: Ibb38c7892326907431e54b7c7dbd7b912b59ccb9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4397707
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>