[RPP] Use source maps to resolve script location when recording is fresh
As of https://crrev.com/c/4358379 the script location is determined
using its URL only. This was done to prevent script ids from a trace
file being matched with scripts of the inspected target, when the file
and the target are unrelated. However that also meant that we lost the
ability to resolve script files using source maps in cases where this
was possible, for example on fresh recordings.

This CL fixes this issue by checking if the current timeline comes from
a fresh recording. If so, it attemps to resolve the script location as
it did before, else it defaults to using the URL.

Bug: 1430444
Change-Id: I1ee640c51d8604976e92cd7393421622c9d62df9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4397719
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>