Add util to adjust event timestamp by previous navigation
On the timeline grid, we adjust times based on the previous navigation.
This means if there are multiple navigations, the times at the top are
reset on each one.

For events like LCP, we can also adjust their timestamp by using the
navigationID parameter to locate the navigation and therefore find its
time. However for some events, like DCL, there is no navigation ID, so
we cannot easily do this. This means if you select a DCL event, the
timestamp you see in the UI may not quite align with the time on the
grid.

This CL introduces a helper function that tries its absolute best to
determine which navigation to use to adjust the time. It will try:

1. The navigationID
2. The frameID, from which we can find the navigation closest to that
   event for that frame.
3. Or, we fallback to event.ts - minBounds, which is our default.

It makes use of the new trace engine to do this, which has most of the
helpers we need in place. Once more of the UI is migrated, we can update
this function to not take in the old data model or SDK Events at all,
and likely move it into the new engine. But for now given it supports
both, I think it should live in TimelineUIUtils.

Bug: 1427403
Change-Id: Ia097958133a5bf25945a12b6e3e9fcf40df5abb1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4402894
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>