Key Payload events by their payload in TimelineData
Now we have some tracks using SDK events, and others using
TraceEventData events, we are hitting a situation where we try to look
up the TimelineData instance for a PayloadEvent, but if we try to do the
same for the same event represented as TraceEventData, it will not be
found.

The solution is to key the data based on the raw payload, which is an
object every SDK.TracingModel.PayloadEvent has, and is the same object
that the new engine represents the event as. These are created once when
the trace is recorded or imported, and never changed.

For ConstructedEvents, the behaviour remains the same, and we do not key
them by anything other than the ConstructedEvent instance, as by
definition these events do not have any raw representation.

Bug: 1428522
Change-Id: I55f8420f74dd8a6054f295c9cea0392536ed7726
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4401937
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>