Move Performance Panel export functionality to new engine
This CL moves the save functionality of the Perf Panel over to the new
engine, skipping the old BackingStorage approach and using a
TextEncoder.

I think that we will now be able to entirely remove the backing storage
code, but I want to do that in its own CL, so I have limited the scope
of this one to leave it in place.

This CL also updates our saved trace files so they follow the
{traceEvents, metadata} format.

Bug: 1428842, 1406847
Change-Id: Id3861cb412f6c6b3e5bcef6c7cf578c3e29eb860
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4381791
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>