[console] Allow evaluation of incomplete JavaScript expressions.
Prior to this change, it was basically impossible to force evaluation of
an incomplete JavaScript expression in the Console (e.g. a for loop with
missing closing brace). This is counter intuitive in various situations
and there should be an escape hatch for developers to actually see the
syntax error.

Therefore this CL introduces two ways to evaluate incomplete JavaScript
expressions:

1. Pressing Ctrl+Enter will always send any (non-empty) Console prompt
   text to V8, without checking whether the string is complete.
2. When the cursor is not at the end of the input, Enter will also
   send the text straight to V8; Shift+Enter can still be used to
   insert newlines. This is consistent with the Console behavior found
   in Firefox DevTools.

Fixed: chromium:1314700
Change-Id: Ie5be6f68aa99c37bb964066480c7ae2990f259f4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4414486
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>