Use end events to set duration of Interactions correctly
When we migrated the interactions track to the new engine, I didn't spot
that the end times of the events were slightly off. I dug into this and
discovered that the UserInteractionsHandler was calculating the duration
incorrectly. Rather than rely on a seemingly rounded `duration` key that
is in the EventTiming start event, we instead have to then pair that
start event with the end event (we can use the ID field) and calculate
the duration based on that.

Additionally, I took this opportunity to make the
UserInteractionsHandler return SyntheticInteractionEvent types, rather
than modify the raw trace event that it is given. In the appender we
push the start events onto the flame chart, but we can then look up the
synthetic event from that given start event.

Fixed: 1432917, 1432585
Change-Id: Iad81939379fe6acd2c5142b71ca63450b1d40f22
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4423190
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>