Calculate nested interactions
For interactions where there are a lot in a short space of time, we want
to hide all but the longest in that block. To do this, we search for
interactions that have the same end time, and then only keep the longest
one in that set.

Note that this CL does not impact the actual track rendering yet, I will
do that in a follow-up, but this CL does all the changes in the handler
to use in the InteractionsTrackAppender.

Bug: 1432510
Change-Id: I1882d9b76f971766b5a0405b39897f55dc43aa9e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4424938
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>