Add helper to calculate event timings
I have started diving into the attached bug, looking at how we can avoid
having to jump from a PayloadEvent to TraceEventData or vice-versa, and
often the only reason we do this is because code that determines the
times of an event expects one or the other. To avoid this, this CL
creates a helper that can return times in Milliseconds both for old
events and new events.

My hope here is that during the migration we can add helpers here to
calculate information, and eventually these can be removed as we stop
depending on the SDK events.

Bug: 1431166
Change-Id: Ic76e13f23082cd5bc51701e0e2d8b5d9881892d4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4431867
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>