Extract TimelineSelection into its own class
So that it can be easily tested. I also refactored a little to make the
constructor private and remove the need to set the type explicitly, as
we can set the type by reading the type of object provided.

Bug: none
Change-Id: I078d652e1b99034083230735ae28a4d7500230c3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4432231
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>