Show INP warning when selecting a long interaction
This CL updates the UI to show a warning when an interaction is over
200ms in duration.

Working on this feature, I discovered a bug where because we were adding
just the beginEvent to the FlameChart, we would not correctly have all
the data available for the UI. So now we push synthetic events onto the
track (which we can do because it extends TraceEventData). This means
when we call rawPayload() on the constructed event (which comes from
CompatabilityTrackAppender#getLegacyEvent), we get the full synthetic
event, and then we can read the data accordingly.

This is a bit messy that the path for when a user selects an event from
an appender track is:

1. User clicks event.
2. We run CompatabilityTrackAppender#getLegacyEvent() to create the SDK
   event.
3. For interactions, we then call .rawPayload() to get the
   SyntheticInteractionEvent back.

Once crbug.com/1433692 is done, we can hopefully tidy this up a lot as
we continue to remove SDK.TracingModel.Event references.

I also think it is messy that we have to, in UI Utils, perform the
calculation to determine if an event is a long duration. I have logged
crbug.com/1434594 to explore a better solution where we can calculate
this data once and access it from both the old engine and the new.

Bug: 1432512
Change-Id: I0c934767a5a41df981967ed4c2dc5d031df5a2eb
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4445940
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>