Unify handling of empty space clicks in Styles pane
Before, it was causing the blank property to be added at the end
for the consecutive clicks to the selector container. I've unified
handling of empty space clicks and changed the logic to add an
empty property to the start of the declaration block if the user
clicked on "selector-container" or "styles-section-subtitle"

In addition to that, added a comment on why we needed
`this.willCauseCancelEditing` property.

After this change, also don't need to set `this.willCauseCancelEditing`
to be false since it will always be checked in `onMouseDown` anyways, before `click` is handled.

Fixed: 1433537
Change-Id: I1a32b6deedf7ec07dc4222b730f10fb8ab90c92a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4466058
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>