Keep bounce-tracking issues alive after non-user-initiated navigation
Bounce-tracking issues can be emitted before an automatic redirect to
a different URL. Normally this navigation would cause the issue to be
discarded. In order for users to be able to see the issue, we keep it
alive until a user-initiated navigation occurs.

CDP's `Network.requestWillBeSent` contains an optional boolean
`hasUserGesture`, indicating whether the request was initiated by a
user gesture. This CL stores this value in the `NetworkRequest`.

The `IssuesManager` discards issues upon navigation. For
`BounceTrackingIssues`, we look up the corresponding request, and if
it was not initiated by a user gesture, the issue is not discarded.

Bug: 1416548
Change-Id: I5dbcb3838da19ea47e37618269c079fbababdc6b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4466087
Commit-Queue: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Auto-Submit: Wolfgang Beyer <wolfi@chromium.org>