Remove TODOs about Layer.requestCompositingReasonIds()
The TODOs were added in crrev.com/c/4144484 based on the plan of
deprecating compositingReasonIds. However, based on the discussions
in crrev.com/c/4457607, we want to keep both compositingReasons
and compositingReasonIds for the flexibility of the protocol to
support future localization of the strings.

Bug: 1432072
Change-Id: I9c94c3e89ba2dc445f64e354ed3ac8f25fd406a7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4455745
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>