Add wiggly lines for console error messages for non-js resources
The primary goal of this CL is to enable pointing to the source of failed stylesheet loads. The implementation goes beyond that by adding wiggly lines to all console errors and warning that are attributed to a source line. This saves us the complexity of having to parse out stylesheet-related console messages.

Bug: 1440626
Change-Id: Ia37fb2998b2d9fa8ee1c0a7034c8865fed02d4b5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4458651
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>