Show popover when hovering on CSS variable
Currently, the CSS variable value on hover will be shown as a title. But, waiting for the title is a bit too long, and many people don't know that they could get the value by hovering.
Also, although you can jump to the definition to get the value by clicking on it, it is still very inconvenient.

This commit uses popover instead of title to show the CSS variable value on hover. On top of that, we can extend more functions later.

Previously when the cursor was located outside the variables of an expression, the value of the entire expression would display.
The following commit highlights the part of the value in effect， so the tooltip on the whole expression has been removed.
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3983512

Video: https://imgur.com/a/aSGAvNJ

Bug: 1380779
Change-Id: I16faf4d2139486deef390f75b1cd08009f5b4fe0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3992726
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>