Reschedule showing popover on mouse move
We want to show popovers only when the user is
keeping the mouse in the same position for some
time. So, in this CL, we reset showing popover timer
when the user moves the mouse even the mouse is
inside the scheduled content.

Fixed: 1441046
Change-Id: I9153e60385814d67c57bcb5df9f3f9dbef1f908e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4489604
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Changhao Han <changhaohan@chromium.org>
Auto-Submit: Ergün Erdoğmuş <ergunsh@chromium.org>