[elements] Delay MDN documentation fetching until popover is shown
This CL removes the eager fetching of the MD documentation JSON when
they styles sidebar pane is first created and instead delays it until
the first time the popover is shown.

This is to reduce unnecessary fetches to our remote server in for
DevTools session where the popover is actually never shown.

R=changhaohan@chromium.org

Bug: 1401107
Change-Id: Id7cc1b86bc8e72e1e181fa45e3a77783d72f6500
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4496389
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>