[sources] Remove CharacterIdMap usage from NavigatorView
This CL removes the usage of PathEncoder from the NavigatorView when
calculating title labels for filesystem mapped directories. This is
because the CharacterIdMap has a limit of 65k "path segment tokens".
We are unlikely to hit that limit in this particular case, but since
it's the last remaining use site of PathEncoder we also migrate it
away and use the path segments themselves directly in the index Trie.

R=bmeurer@chromium.org

Bug: 1436039
Change-Id: I8301a49d9af45a56f60220ad00426cc0d84a7c59
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4498705
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>