[RPP] Extract events for tree views in appender system
The tree views in the details pane are fed using the events of the
selected track. The trees show the events in the track that can be
shaped as a tree. This is always the case for sync events, but for async events a check needs to be done. Events can be shaped as a tree if  there is *not* a pair of async events e1 and e2 where:

e1.startTime <= e2.startTime && e1.endTime > e2.startTime && e1.endTime > e2.endTime. or, graphically:

   |------- e1 ------|
        |------- e2 --------|

So that we can use data from the new engine in the tree views, this CL
adds some helpers changes to the CompatibilityTrackAppender to keep
record of the tree events for each track:

1. eventsInTrack: Calculates all the events shown in a track.

2. canBuildTreeFromEvents: Given a list of events, determines if a treecan be built with them.

3. eventsForTreeView: given a track appender name, returns the events to
   be shown in the tree views. This makes use of the 2 previous
   functions to determine which events need to be returned (the result
   of 1. if 2. returns true, else the sync events in the track).

4. groupEventsForTreeView: given a flame chart group, returns the events
   for tree views in the track related to that group. When a user clicks
   a track in the flame chart, a FlameChar.group is passed to the
   appender layer. This Method translates the group into an appender and
   returns the result of 3.

In an upcoming CL, the details view artifacts will be updated to use
trace events, instead of using the legacy track definition to obtain the
events to be used for tree views (see [1]). When that is in place we
will be able to plug the new engine events using the changes in this CL.

[1] https://docs.google.com/document/d/1yhOwUj0abz6gkc5iDd0YwweO68P0XxX0etgnd0XI54M/edit#heading=h.pcawn2wj7knb

Bug: 1433692
Change-Id: Ic1113daf538670ebb433565170ad6604bf0d331a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4481319
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>