[uma] Record whether users look at minified, source mapped or plain JS
This CL increases the resolution of 'text/javascript' files in the
Sources and Network panel. Specifically we now distinguish between
minified, source mapped and JS as authored ("plain").

This requires us to delay metrics collection slightly until the
content of a UISourceCodeFrame is available. Note that slightly
changes the semantic of the "SourcesPanelFileOpened" metric:
We now record the metric only if the user actually opened the
source file tab and looked at the content. Until now we already
recorded the metric when we created the tab.

R=jarin@chromium.org

Bug: None
Change-Id: I63493d9df8accf4672b4c37ab5d202f4cb23e7ad
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4508851
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>