[RPP] Remove references to legacy track in details views artifacts
This is part of the implementation of the proposal doc to use new engine
data in the details view [1]. Instead of referencing track objects, we
reference event objects, which is ultimately what we need. This way we
can use data from the new engine in the details views.

[1] https://docs.google.com/document/d/1yhOwUj0abz6gkc5iDd0YwweO68P0XxX0etgnd0XI54M/edit#heading=h.pcawn2wj7knb

Bug: 1433692
Change-Id: I28a3a53908a19647b5011a94f849ede7b1a9fe9f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4481135
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>