[elements] Hide search on navigation/reload
When redrawing the elements panel, e.g., on reload and navigation, this
CL closes the search toolbar in case it's open instead of only resetting
it. The user can bring it back with single ctrl+f and it will still
contain the previous search. This cleanly updates the search results.

Fixed: 1189064
Change-Id: I18f4d81af604279e8cf3c2ec0b6b245988abb2c6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4506959
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>