[sources] Clear execution line deco when disposing DebuggerPlugin
This CL fixes a bug that would show the "paused" line decoration in
the CM editor after a reload even though the page is not paused.

This is because the old DebuggerPlugin that gets disposed doesn't
clean up after itself properly. This might have worked before
https://crrev.com/c/3970719 has landed: Now we reset the `this.editor`
to undefined so the "setExecutionLine" method bails out early if it's
called on an already disposed DebuggerPlugin.

The fix is to clear the decoration explicitly when a DebuggerPlugin
instance is disposed.

R=jarin@chromium.org

Fixed: 1442968
Change-Id: I9e8287e97d7d2e0a6678be44f3568e5ba8d54b11
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4513702
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>