[console] Use source map resolved variables for autocomplete hints
This CL integrates the existing "source map variable resolving"
mechanism of the scopes view with console autocomplete suggestions.

Implementation-wise this is rather straight-forward as the
"SourceMapScopes" module already implements a RemoteObject that does
the mapping of scope variables for us, we just have to use it :)

Drive-by: Rename 'frames' to 'scopes', since that is the thing we
map over.

R=jarin@chromium.org

Bug: 1444349
Change-Id: I4b95108f89a2d4db4f49c86cdc6521637e52c2e1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4524939
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>