Set all JavaScript breakpoints eagerly
This patch adds a behind-a-flag functionality to set breakpoints
eagerly. In practice, this means setting all the script
(i.e., non-source-mapped) breakpoints from the storage in
the breakpoint manager when the breakpoint manager is constructed.
When the debugger attaches to a target the breakpoint manager
will automatically set the breakpoints in the target.

Bug: 1442232
Change-Id: Ibd81c28988ad65869d4cbf24bd05494533317ffd
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4524942
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Jaroslav Sevcik <jarin@chromium.org>