Improve click box for legacy DataGrid
The DataGrid code that checks if the user is expanding a row was using a
10px wide box to listen for clicks. The icon itself is 14px wide, so it
was very easy to click on the arrow but actually quick outside the
bounds that the code checked. This CL updates the width (and makes it a
private variable as a drive-by) to fix the click box.

Bug: 1447043
Change-Id: I7ec3076f69dc024f17ef4bcb772ee406ab8739f9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4549147
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>