[metrics] Report DevTools.Workspaces.PopulateWallClocktime.
This adds the logic to report the time to initially populate a project
folder for the Workspaces feature. The FileSystemWorkspaceBinding also
deals with the snippets and the overrides features, which we don't care
about here, so we restrict this to just Workspaces ('filesystem').

Bug: chromium:1447968
Back-end: https://crrev.com/c/4555256
Change-Id: Ib721145ea3a0ffa39bef1fdd5ac9dc719b2d26fa
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4553660
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>