Remove sections from the shared storage metadata view.
There second section is not a section at all and is used as a header for
the bottom pane in the split view. We don't label similar panes for
other storage types and even if we want to, we should abuse report
sections for this.

Bug: 1448011
Change-Id: I4be10327f50651c8fe352c9f07831a922a46778e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4556928
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Danil Somsikov <dsv@chromium.org>
Auto-Submit: Danil Somsikov <dsv@chromium.org>