[DIPS] Re-render Bounce Tracking Mitigations panel before async call
This CL ensures that the BTM panel makes it clear when mitigations are
running in the background by not waiting for their results to return
before re-rendering the panel. This way, the "Force Run" button is
disabled and a loading message is displayed while mitigations run.

http://screencast/cast/NTAzOTc0NDc1Mzg2MDYwOHw5NTE0NTI5MS1lNQ

Bug: 1449920
Change-Id: I1dfc28bec38a49560bd408aa695d754f5400d02a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4577138
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Joshua Hood <jdh@chromium.org>