[web tests] Remove simple test runners from globalThis
This CL removes the test runners, that are defined in only one file,
from globalThis.

The rest of the test runners are split across multiple files where
each file modifies the same global. The change to fix those is not
as straight-forward.

This CL is blocked on https://crrev.com/c/4584291 which adds some
missing imports to web tests.

R=dsv@chromium.org

Bug: 1442410
Change-Id: I7ffa92854bdac882350d979ef47ced74a73cff30
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4584668
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>