Avoid relative paths in ESLint config
If ESLint is run directly on a file nested in a sub directory, it may
not find the right TSConfig file to use (I see this happen in my editor
a lot where the editor's ESLint plugin cannot locate the file as the
path is wrong). Swapping the path to be an absolute path fixes this.

Bug: none
Change-Id: I35aa687df2bebddfd28ced91c4cd96ece0cce7c1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4584590
Commit-Queue: Kateryna Prokopenko <kprokopenko@chromium.org>
Reviewed-by: Kateryna Prokopenko <kprokopenko@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>