Enable canvas context tracking for PerfUI
Enables the canvas context tracking rule, but only for the PerfUI
folder, because it only needs to apply when there is a bunch of canvas
use.

We do not have any violations of this rule currently, but I added one
manually and have confirmed that the rule does detect it and fail the
build.

Fixed: 1450569
Change-Id: I59f545beb3d0354f9ae3f9f864af3314bb27baed
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4588912
Commit-Queue: Nancy Li <nancyly@chromium.org>
Reviewed-by: Nancy Li <nancyly@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>