Prevent screen reader from announcing checkbox label twice
Due to duplicate labeling from the aria-label of the checkbox and from the label element's for attribute, screen readers announce the checkbox label twice. This patch prevents that by removing the aria-label on the checkbox and instead setting a title attribute.

Note the for attribute on the checkbox label was last added in this CL - https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2077666
to address a regression introduced here - https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2062603.

There's a bit of friction with this change because it requires updating existing tests that depend on the checkbox aria-label selector. Future tests would also need to align with this change.

There was an initial attempt to land a similar change but then that got reverted in here - https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3948895

Nonetheless, having both an aria-label on the checkbox and the for attribute on the label will cause the duplicate screen reader announcement.


Bug: 1414952
Change-Id: If4342b860d07b0a9494b90f4d15cb99688cb9d6e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4591249
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Sylvester Elorm Coch <elormcoch@microsoft.com>