Fix multiple trace stops in NetworkPanel
I noticed when working on the NetworkPanel that if screenshots are
enabled and we start tracing, we will get an error about trying to stop
the trace manager twice.

The issue is that we bind to an event, but this callback can be fired
multiple times. So if we get the event a second time, we first cancel
the timeout, to avoid queueing multiple stops.

This wasn't causing a user visible bug, ultimately stop()ing twice is a
no-op, but it tidies up a console error :)

Bug: none
Change-Id: If7b16d67b7517c3e12b54f91902496aba98f5ae3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4599157
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>