Remove async nature of SDK.TracingModel.ObjectSnapshot
The ObjectSnapshot class had this entire concept of a backing storage
but by looking through the code I confirmed that this was never used.
Therefore we can simplify the code a lot by removing that and then
making the method to return the data not async.

There is more that can be done here: ultimately I think given this
architecture we do not even need the ObjectSnapshot class and can
represet it as a normal PayloadEvent, but I want to work in small steps
and avoid large CLs. I have left some TODOs which I will also address in
follow-up CLs.

Bug: 1453234
Change-Id: Idf85d8026a691e457af47c6dd2f53e998a80e062
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4604651
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>