[RPP] Wait for CPUProfilerModel to exist before adding listener to it
Before this CL the event listener that loads a profile via
console.profile() was added using TargetManager::addModelListener. This
is problematic because there is no guarantee that the model exists by
the time the event listener is added (f.e. if the target hasn't been
created). As a consequence, using the console.profile() API wouldn't
work under this condition: Open Node DevTools with the Performance panel
as the initially selected tab.

This CL fixes this race condition by adding the event listnerer to the
model only when the model has been created.

Fixed: 1457197
Change-Id: I3143d02c197b7b9be17b2b5ba8596ab20c65d148
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4637161
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>