Prevent issues from being discarded incorrectly on prerender- activation
Issues can be generated for a prerendered page. When such a
prerendered page is activated, this CL prevents the corresponding
issues from being discarded. The result being that the issues are
shown to the user after the prerender-activation.

Bug: 1393057
Change-Id: I262374e5705a56be0822b6eaf683fb8cc87aef62
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4650556
Auto-Submit: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Danil Somsikov <dsv@chromium.org>