Fix duplicate top layer badges
Currently we sometimes show duplicate top-layer badges for a single
element, because we run updateTopLayer for each DOMDocument (iframes),
and for same-root iframes, where dom IDs are shared and DOMModels are
the same, it's impossible to distinguish which frame the top layer
belongs to. This CL adds a DOMDocument to the TopLayerContainer, so that
every insertion is checked against this DOMDocument, to ensure that we
only add top layer badge when the top layer element belongs to the
current TopLayerContainer's DOMDocument (DOMDocument is unique for each
iframe, regardless of if its same-siteness).

Fixed: 1359978
Change-Id: I35aab5745bfc053686db4b2c75db2a9d26630df8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4650553
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Changhao Han <changhaohan@chromium.org>