[unit] Don't access an already-destroyed promise type in a promse.then
In the instance of the linked bug, a Promise was stored in a
Promise.race. If the promise completes after the afterEach hook that
stops tracking, it couldn't access the original Promise type anymore
because that was flushed, which meant it couldn't call that type's
`then`.

Fixed: 1451331
Change-Id: I891467523a9ac53936377d9eb0f8e7b742bde7f1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4660439
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>
Auto-Submit: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>