Remove validation on blur for ListWidget input field
Validating on input is sufficient. Running validation on blur
causes the validation to run even when the user clicks on cancel,
which causes a flash of an error message to be displayed before
the editor is removed.

R=leese@chromium.org

Fixed:1461986
Change-Id: Ia0d0d4a27a1870f85722544d412ff17c32f7134d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4660371
Reviewed-by: Eric Leese <leese@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>
Commit-Queue: Yang Guo <yangguo@chromium.org>
Auto-Submit: Yang Guo <yangguo@chromium.org>