Break JS tracking stack when top level event is found
As we are building JS stacks, we have the possibility of nesting trace
events inside JS frames, which makes sense in some cases, like when a
styles recalc is triggered by a script, but doesn't in other cases, like
when a RunMicrotasks follows JS samples. In the latter case, the JS
stack needs to be restarted since RunTasks and RunMicrotasks are top
level trace events.

Bug: 1455214
Change-Id: Idc64bd7912a0994ce906bb3d14ace65b29b9dbd8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4665434
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>