Disabled navigation on resize/move animations when animations and elements view are in the same location
Issue:
When resizing or moving animations in the animations view, the view immediately navigates to the elements view. This becomes an issue when the elements view and animations view are in the same location (e.g panel or drawer).  The user is no longer able to resize or move the animation since the view disappears.

Changes:
Added a check to not navigate to the node in the elements view if the animations view and elements view are in the same location.

Before and After:
https://imgur.com/a/i4ns8Ad

Bug: 1459043
Change-Id: Ifba9b4c256f0ef4760a25d0993e19c32b8d66f1f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4653280
Commit-Queue: Jimmy Seto <jimmyseto@microsoft.com>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>