Remove getFirstFCPTimestampFromModelData from handler
This function was used in Insights to position the playhead. We don't
have any need for this in the Perf Panel (yet) and we can easily bring
this back in the future. For now let's remove it to avoid confusion by
having a function that we never call.

Bug: none
Change-Id: I13d279a441db668462654b0cd5b10b6d1ffe71c0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4667395
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>