[scopes] Don't look for function scope when collecting identifiers
With https://crrev.com/c/4650348 we parse the full script when
collecting identifiers for a script. Previously, we searched for a
surrounding function scope and used that as the boundary for parsing.

We still used the function scope to get the underlying script from,
but this not strictly necessary since we can get the script also from
the actual scope we are looking at.

This CL cleans up the code in NameResolver by removing the function
scope search alltogether. This has the nice benefit that we now
also handle variable renaming on the module level. This did not work
previously as we didn't find a function scope surrounding the
module scope.

R=jarin@chromium.org

Bug: 1444349
Change-Id: Ibaacdc4f7f1478023e4e5062df65d026e9b0ab80
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4667393
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>