Restructure loading trace files in tests
I was getting a bit lost with how we load trace files in tests; there
are many different methods - some are cached, some are not, and there is
a lot of overlap.

This CL changes that by creating TraceLoader with exposes a few methods:

- fileContents: load the raw contents of a trace file
- rawEvents: load the raw events array of a trace file
- traceEngine: load the file and run it through the new trace engine
  only
- allModels: load the file and generate all the models

These methods all use a cache to avoid parsing any one file more than
once, and we also cache the creation of the models.

Bug: none
Change-Id: I8fb98903efc2e18dda9f451cc1fc6a40554450f1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4675392
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>