[Protocol Monitor] Clean up code for JSONEditor.ts file
The functional part of the CDP command editor is mainly done, but the code could be further improved with some refactoring. The changes include the definition of new interfaces for parameters, the removal of unneeded functions and the transformation of if/else statements into switch blocks.

Bug: 1462919
Change-Id: I2c5829fbfa46525232a64dc18c80bc4ad44aabd7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4675886
Reviewed-by: Nancy Li <nancyly@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Hadrien Jaubert <hadrienjaubert@google.com>