Add screenshot test for TimelineOverview
Before we begin to migrate it. Note that some of the styles copied into
the component example are required to make it look just like it does in
the Performance Panel. In a follow-up CL I will try to move these styles
into the Overview components themselves if possible so that we can avoid
the duplication.

Bug: 1464206
Change-Id: I24952b9b07800e19bbfc16b725126863f88fef29
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4679608
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>