Re-enable data grid controller tests
The error that initially caused these tests to be disabled was:

```
  ProtocolError: Protocol error (Runtime.evaluate): Promise was collected
      at new Callback (node_modules/puppeteer-core/src/common/Connection.ts:66:12)
      at CallbackRegistry.create (node_modules/puppeteer-core/src/common/Connection.ts:127:22)
      at Connection._rawSend (node_modules/puppeteer-core/src/common/Connection.ts:278:22)
      at CDPSessionImpl.send (node_modules/puppeteer-core/src/common/Connection.ts:537:29)
      at ExecutionContext._ExecutionContext_evaluate (node_modules/puppeteer-core/src/common/ExecutionContext.ts:288:10)
      at ExecutionContext.evaluate (node_modules/puppeteer-core/src/common/ExecutionContext.ts:189:32)
      at IsolatedWorld.evaluate (node_modules/puppeteer-core/src/common/IsolatedWorld.ts:210:20)
      at Context.afterEach (test/conductor/mocha_hooks.ts:146:26)
```

This seems to be Pptr related and not a direct issue with these tests
(it also occurs in the global afterEach block, not in the data grid
tests themselves)

Since these tests were disabled we have rolled new versions of Pptr into
DevTools, and I cannot recreate the crash locally, so I think we should
re-enable and see if they are stable. If not, we can then dive properly
into hunting down the issue.

Bug: 1442175
Change-Id: I640fbc8d1345a32b16ba576cc80e2b69311c8120
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4683631
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>