[sources] Automatically reveal viewed file in the Sources sidebar.
This changes the default for the "Automatically reveal files in sidebar"
setting to true, to align DevTools' default behavior with the default
found in IDEs like VSCode. This is to address the discoverability
problem with the setting, and address the requests of developers who'd
love to have this feature (but didn't realize there's already a
setting). For developers who explicitly don't like this, which we
assume is the minority by far, they are likely able to locate the
setting to turn it off.

Fixed: chromium:1336599
Bug: chromium:1301265
Change-Id: Idb585bf908d8ebdf4b95e8b99c7b04977ff396ca
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4684349
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>