Pass traceParsedData into HistoryManager for preview
When we store a new trace the history manager constructs a preview for
that trace using the TimelineEventOverview components. Now that we are
migrating those to the new engine, we need to pass the new engine's data
in when storing a history entry so that it can construct the preview.

To ensure this is working, I updated some of the components to store a
reference to the new engine. They are not using this yet, but they will
in follow-up CLs.

Bug: 1464206
Change-Id: I0c42236c41c3a3eab14eb83ed76ae5c8b65a4770
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4685564
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>