[breakpoints] Port inline script breakpoint web test to e2e
The test replaces
http/tests/devtools/sources/debugger/debug-inlined-scripts.js
which isn't working for a year now due to the CodeMirror upgrade.

Note that we don't replace the full range of checks: When pausing
in inline scripts on load we don't always have the full HTML
document ready due to a race in the backend.

R=jarin@chromium.org

Bug: None
Change-Id: Ieadfd538399c8990a338184377f6a68addfe56e7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4700881
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>