Add WarningsHandler
The old data layer implements a bunch of warnings that we do not have in
the new engine.

Previously in Performance Insights we implemented these per handler, but
I think I want to instead have a single handler dedicated to gathering
up anything we might want to consider a basic warning.

In future CLs I will add warning calculations to the handler, and then
begin to use the warnings across the codebase, removing the old warnings
code in favour of this.

See the attached bug for the initial set of warnings we are going to
support.

Bug: 1465383
Change-Id: I507637c00e85f68f7bab5e1e5e9a891a0c5582e2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4701063
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>