Increase the default DevTools unittest timeout
More and more of our tests use `describeWithEnvironment` and the
environment is naturally increasing over time and getting larger. We are
seeing the odd flake where this can take longer than expected. This CL
is to see if it is crashing, or if a slightly higher timeout will give
the bots some room and increase the reliability of the tests.

Bug: none
Change-Id: I7a49f6ad50912175076470f2b8efd3c0b229f6a2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4724886
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>