Use threshold logic from UserInteractions handler
Long term I would like to rework how we generate these errors - and
remove the enum that represents warnings and move to using the new
engine's details, but one step at a time! This change removes the
duplicated logic for determining a long interaction and relies on the
handler.

Bug: 1467192
Change-Id: Id23be0759b9c1fd2548b407861af23c81f03f48a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4724884
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>