[scope] Fix 'value unavailable' for source mapped variables
This CL fixes a small bug when we calculate the properties for a
remote scope object where we source map the scope's variables.

The current code filters out properties where the value is `undefined`.
This breaks the "value unavailable" experiment, where we report
unavailable variables (e.g. TDZ) with no value and no getter.

This CL fixes this by removing the undefined check. Furthermore the
CL will now also create a proper copy of the remote property and
copy getters and setters. This does not affect the scope view
directly, as V8 does not put any accessors on scope objects, it's
cleaner nonetheless. We should either clone the whole remote property
properly, or assert that no getters/setters are present. This CL
chooses the former.

R=jarin@chromium.org

Bug: 1328681
Change-Id: I72ea76b3ff9d6e270c3f95e20deb49a2e610b372
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4728043
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>