[Accessibility] Narrator does not announce if menu from tabbed pane is opened or closed
Issue:
Narrator doesn't announce if the menu is expanded or collapsed for the '>>' button on the toolbar.

Changes:
Added logic to detect and set aria-expanded to true or false for the '>>' button on the toolbar.

Before (Not Expanded):
https://imgur.com/plyQmyy
After (Not Expanded):
https://imgur.com/eV6SOMa

Before (Expanded):
https://imgur.com/Qd1EmlL
After (Expanded):
https://imgur.com/gh2OA7X

Bug: 1468631
Change-Id: I765b92725ffde25a5405d91974acaba8e3f45217
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4729777
Commit-Queue: Jimmy Seto <jimmyseto@microsoft.com>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>