Use all handlers in basic component doc example of the RPP
When this example is loaded, the perf panel is loaded with all the
hanlders enabled, instead of only those enabled during the migration.
This way we play with how the renderer tracks coming from the new engine
look as we migrate to the new engine, without actually shipping the new
tracks in the full DevTools version of the panel.

Bug: 1428024
Change-Id: I79a196447486790ef4f5f4b22799bbda37051db5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4729647
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>