[RPP] Mark renderer and samples handlers as optional during migration
With https://crrev.com/c/4729647 in place, we are now executing all
handlers when running from the component example. However, our types do
not reflect that yet.

This CL updates the migration types to mark the renderer and samples
handlers as optional properties, so that we have type support for them
when trying to render thread tracks with the new engine. Otherwise, they
would not exists from TS perspective and we won't be able to use their
data.

Bug: 1428024
Change-Id: I2e105b9f4ab526e6c0d448854cdc39a38fa7cae2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4733292
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>