[search] Add 'lineSegmentForMatch' method
This CL adds a helper method 'lineSegmentForMatch' for an upcoming
search panel change. We land the helper separately upfront because we
have to touch a bunch of BUILD.gn files for the new unit tests and
this would make a single large CL rather noisey.

We plan to change the search panel to include one result entry per
match, rather then one result entry per line where we highlight
multiple matches on the same line (see https://crbug.com/1468875).

Given a line and a range that describes the search match, this CL
implements the logic that creates a string view of the line where
the match is "centered". This is necessary to show search results
in minified code, where we usually only have a single line with
hundres of thousands of characters. With the helper we'll be able
to show the area around a match for each match separately.

R=dsv@chromium.org

Bug: 1468875
Change-Id: I11c80957d8d8a1b9a277c80d8f0dbdcd221480d3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4738595
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>