Add Storage Buckets to Application panel
Adds the Storage Buckets menu to the sidebar of the Application panel.
The Storage Buckets menu is a tree which contains a list of all the
storage buckets for the page. Under each storage bucket will be  a list
of supported storage APIs. To use, this must be enabled in the
Devtools' experiments settings menu.

Bug: 1406017
Change-Id: Iecde698d347aa9c3fe650c410804fbe49dfd5d3c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4669408
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Sofia Emelianova <semel@chromium.org>
Commit-Queue: Nathan Memmott <memmott@chromium.org>