Ensure Perf panel recovers from an error when tracing
This CL fixes an issue where DevTools could get into a broken state if
the following happened:

1. User tries to record
2. Something in the backend goes wrong for whatever reason.
3. User tries to record again.

At this point, the user would see "Tracing already started". This error
comes from the backend and happens because in step (2) above, we started
tracing, but then errored, and never stopped tracing. Therefore at this
point the frontend thinks we are not recording, but in the backend we
are.

The fix here is to introduce a reset() method to the TracingManager
which will invoke the CDP Tracing.stop method if we have an active
client. This ensures that after an error we tell the backend to stop
tracing.

Fixed: 1465261
Change-Id: Id5f9ec14ef6ce3b16519bb956f625c0a65891570
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4753195
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>