PreloadingDevTools: Remove 'Activate' button temporarily
Currently, 'Activate' button for prerendering not working correctly and
always fails to activate due to parameter mismatch. We need to discuss
to fix this. So, this CL temporarily removes the button for M117 to
reduce confusion.

For more details, see https://crbug.com/1467025#c6.

Bug: 1467025
Change-Id: I7a4edf01917b2a1d59936443286ce937b577f05c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4756135
Commit-Queue: Ken Okada <kenoss@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>