Add border around "Total" on PerfUI flamechart
This CL adds a border around the total box on the pie chart, otherwise
it looks like there isn't one and it is broken. It is also less obvious
for users to understand they can click on it and select it.

Fixed: 1470147
Change-Id: I0e15ce99651261524ac5c7a4dcffacb5e47d8eb6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4753201
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>
Reviewed-by: Nancy Li <nancyly@chromium.org>