[search] Pass/return UISourceCode directly intead of URLs
This CL updates the "findFilesMatchingSearchRequest" interface method
and it's implementations to filter on UISourceCodes instead of URLs.

This saves the going from UISourceCode -> Url and back again for
ContentProviderBased projects. For file system projects we still
have to do it but we pushed that into the implementation.

The motivation is to cache search results from the backend later on
in "ContentProviderBasedProject#findFilesMatchingSearchRequest" keyed
on the UISourceCode.

R=dsv@chromium.org

Bug: 1468875
Change-Id: I0dbce54da6bcd90521faaa906a6dfdcd9150461e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4762665
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>