[search] Move SearchConfig from panels/search to models/workspace
SearchConfig is an interface with a single implementation in
panels/search. The implementation doesn't need anything from
panels/search so we can make the code a bit easier by removing the
indirection via an interface.

This CL only moves the class but does not yet remove the interface.

R=dsv@chromium.org

Bug: 1468875
Change-Id: I1de0b12196b9f579e7196562a83ac0df25fdeda1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4765267
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>