Re-enable data_grid_controller test
This was disabled a long time ago, but the failure was in the afterEach
step, and does not seem to be specific to this test. Therefore I am
re-enabling it to see if it is passing now or still flakey.

Fixed: 1463394
Change-Id: I1c5734dcd9b61c0379540c4612556c4b22ead94d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4765268
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>