RPP: Minimap Breadcrumbs data structure
We are going to add breadcrumbs to the Performance Panel to aid users as they zoom in and out of the timeline.
This CL adds a data structure to store breadcrumbs. Because breadcrumbs are nested within each other, we represent them as a linked list.

Bug: 1471463
Change-Id: I7d66a1e7f10ba1cc59d74bc037e09b4e38512701
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4760705
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@google.com>