Ensure network data provider only returns main frame navigations
We had a bug (see below) where the timeline in the panel would
unexpectedly reset to 0ms. This is suppoosed to only happen when the
trace has navigations in it, but it was happening on a trace where the
only navigation on the main frame was the initial one.

The NetworkDataProvider was incorrectly returning all navigation events,
and not filtering them to ones only on the main frame, which caused
this bug. The fix here is to filter based on the main frame.

In a follow-up CL I will move this functionality into a helper, because
the main data provider should also use it.

Fixed: 1471447
Change-Id: I8a5d193e2c189cc0980a1ddf590d0d4e638fb2cf
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4768206
Commit-Queue: Nancy Li <nancyly@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Nancy Li <nancyly@chromium.org>