Add timing types to TimelineOverviewCalculator
This CL updates the TimelineOverviewCalculator by moving it into its own
file, and changing its types from `number` to the `MilliSeconds` type.
This is important because the trace engine works primarily in
microseconds, so if we just use the `number` type we could easily get
into a confusing world where we are passing round numbers in different
units.

Bug: 1464206
Change-Id: I439c120c83950c2ca2cfa0c892b9f9065359dba7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4777912
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>