PreloadingDevTools: do not truncate tooltips on urls
How it works:
Cell.title field is used to display tooltips. And if it is not set,
Cell.value (in this case, url) will be truncated to the first 20
characters and be displayed.

Previously, we did not set Cell.title, so the url was truncated.

This CL sets the Cell's title field to the complete url to avoid
truncation.

Before: See the issue's description
After: https://crbug.com/1470614#c4 and https://crbug.com/1470614#c7

Bug: 1470614
Change-Id: Iaf4fe9083cb4b399b87eb57c4b1468facf88b245
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4776365
Reviewed-by: Ken Okada <kenoss@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Lingqi Chi <lingqi@chromium.org>