[search] Introduce 'SearchMatchExact' type and search functions
The existing `SearchMatch` type is better described as a "match hint".
It only describes a line on which matches are very likely to be found.

This CL adds a new "SearchMatchExact" type, that describes a specific
match on a given line. It includes the column (index) where on the
line the match is found as well how long the match actually is.

The CL also adds 2 new functions that mirror "performSearchInContent",
but rather then returning a "match hint" per line, we return one
"SearchMatchExact" instance for each actual match.

Follow-up CLs will update the SearchResult interface and UI to
utilize the new type.

R=dsv@chromium.org

Bug: 1468875
Change-Id: I4f312eb47138478ebded65cc361e0d26ce87d5bf
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4783307
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>