[search] Support 'one search result per match' in the SearchResultsPane
This CL adds a 'matchColumn' and 'matchLength' to the 'SearchResult'
interface. In the future an entry in the result can describe an exact
match. We provide that info in the model (as it's known) so the UI
doesn't have to re-execute the search in that case.

Note that we support both modes in the SearcHResultsPane. Searching
in network headers/urls still only produces one search result entry
where we have to highlight all matches simultanously.

R=jarin@chromium.org

Bug: 1468875
Change-Id: I5cae53a833e85343b5706658ab33d9464e7f74b1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4789167
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>