Add branded types for Timing
This Cl adds Platform.Timing and some helper methods to create and
convert between times.

They look suspiciously similar to the types in the TraceEngine and that
is on purpose. The reason that I haven't updated the TraceEngine types
is that long term the TraceEngine will not live within DevTools but will
be its own repository so it can be easily reused across all our
performance tooling.

However, by making the types in TraceEngine identical to the types in
Platform, this means that this function:

```
function foo(x: Platform.Timing.Seconds)
```

Can take both Platform.Timing.Seconds and TraceEngine.Timing.Seconds and
still be type safe.

I will follow this CL up with some that begin to use these types across
DevTools, particularly in components that are re-used across the
Performance and Network panels, where currently `number` is used and
causes confusion given the network panel works in seconds and the
performance panel uses microseconds.

Bug: 1473917
Change-Id: I3afab3d53520fd10dae16e087f26a9b9cb4fbb29
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4790764
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>