[search] Preserve search results in 'findFilesMatchingSearchRequest'
This CL changes the 'findFilesMatchingSearchRequest' return type from
a list of UISourceCodes to a Map keyed on UISourceCodes.
The ContentProviderBasedProject already does a prelminary search in
the backend as well as in the content (if present) and throws away
the search result.

This CL is a first step to re-use the exact search results by
returning them. The "file based project" implementation
unconditionally omits the search result.

We don't want to re-use the search results from the backend as
that implementation slightly differs so in the end we always
want to search in the frontend.

R=dsv@chromium.org

Bug: 1468875
Change-Id: Id46df819973e1caeb0dcd2b605012de35f60c145
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4771024
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>