PreloadingDevTools: Make selector's ItemId non-null to fix highlight bug
The null type has different meanings in different scopes:
Within `UI.ListModel.ListModel`, the null id can be used to indicate
no items are selected.
Within `PreloadingRuleSetSelector`, the null id is used for the
default filter that selects all rules. (so we treat it as a root
rule now)

When PreloadingRuleSetSelector is interacting with ListModel, the
intention of the null type gets unclear. It is hard to distinguish
whether a null item ID means selecting nothing or selecting the root
rule, then it causes the "All preloads" item cannot be highlighted
as the ListMode thinks users select nothing.

To fix this bug, this CL:
- Introduces a new type of symbol for the root rule.
- Makes PreloadingRuleSetSelector use the new type instead of the null
  type.
- Add another layer that translates the item id to rule set id, and
  uses the converted Id when interacting with PreloadingModel

Before: https://crbug.com/1470621#c5
After: https://crbug.com/1470621#c6

Bug: 1470621
Change-Id: Ifcc709685ac721cd2d59498c2c716c33b16658df
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4813115
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Lingqi Chi <lingqi@chromium.org>
Reviewed-by: Ken Okada <kenoss@chromium.org>