[search] Speed-up search result rendering
Currently, we add search results one at a time to the UI. Concretely
we:

  1) Add a new entry to the tree in the results pane
  2) Update the counter label in the search view

This causes DOM layout and paint, which increase over time as the
number of results grow. As every search result is added individually
this compounds over time.

This CL throttles the UI updates by rendering the search result in
batches. This greatly reduces overall search time when the number
of results is large. E.g. a search yielding ~4 million results in
~2500 files was reduced from ~23 seconds to ~5 seconds.

R=dsv@chromium.org

Bug: 1468875
Change-Id: I20a7f739ec79b65ab1d15f41876efd876cdae591
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4809936
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>