[search] Add unit tests for NetworkSearchScope
This CL adds unit tests for the most important NetworkSearchScope
behaviors:

  * Searching for requests by URL
  * Searching for specific request/response header names/values
  * Searching in the response body

Note that the response body search is slightly broken at the moment
due to wrong usage of the SearchConfig: It uses the user-entered
search query directly instead of the parsed 'queries' array. This
means that multi-part queries and 'file:' prefixes don't work
for the request search at the moment.

Drive-by: Inject the NetworkLog into the search scope instead of
using the global. This allows the test to use a fake.

Drive-by 2: Fix the network search result 'matchLabel' returning
a number instead of a string for response body searches.

R=dsv@chromium.org

Bug: 1468875
Change-Id: I99b9ebf0f3fcc9b606fffd0ea1ba3c97286ca252
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4815467
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>