[RPP] Cache thread by worker id
In order to assign the name of the track for a worker thread in the ThreadAppender, we need to know its worker thread id. This data is available in the events cached by the worker handler but it would be easier to consume if it was stored in key-value pairs.

Bug: 1428024
Change-Id: I4bbd22be52a890c45ae4c2b1ed07de3b31bd1767
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4817929
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>