Introduce TreeManipulator
This CL lands the very first version of the TreeManipulator which is
intended to provide support for various user actions on a flame chart,
including collapsing stacks, merging functions, collapsing recursion,
and so on.

There is plenty more functionality we want to implement (see attached
bug and subsequent design doc); this CL just lands the initial version
and a few tests of the basic functionality that has been implemented.

Long term we see any form of filtering that is possible in the
Performance Panel (event type filters, ignore list, etc) being
implemented via the TreeManipulator. Everything in this CL is just
laying the groundwork and is open to being changed as we expand and use
this class.

Bug: 1469887
Change-Id: Icbfb2f59d9788d70fbba814ccc9fee6ccd1d2fba
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4759866
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Nancy Li <nancyly@chromium.org>