[Performance Panel] Migrate color names inside Timeline to Material 3 and add some small UI changes
This CL concerns really minor UI changes but more names updates.
The visual changes are very minor such as for instance the background color of the timeline-overview-strip-title or the border-color of the paint-view. The "only" visual update is the update of the fillStyle for the TimelineFlameChartNetworkDataProvider.

Bug: 1456706
Change-Id: Ib7d8b20d7076f44f7487273db81757f72a3f5bb4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4822452
Reviewed-by: Kateryna Prokopenko <kprokopenko@chromium.org>
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Hadrien Jaubert <hadrienjaubert@google.com>