prerender: remove properties/capabilities of ResourceTreeModel and ResourceTreeFrame
The old prerendering section was removed in the previous CL [1] because
the same information is available in the new panel's "this page" pane.
[1] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4835675

To fully remove the code related to the old prerendering section,
this CL removes properties/capabilities of ResourceTreeModel and
ResourceTreeFrame since preloading panels is built on top of
PreloadingModel.

Bug:1445219
Change-Id: I2e503a063a1eac0bff1a028acb073dde72107a59
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4840769
Reviewed-by: Ken Okada <kenoss@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Miina Koyama <miinak@google.com>
Reviewed-by: Lingqi Chi <lingqi@chromium.org>