[worker] Replace 'create' with 'createLoader' disposition
'create' is only ever used to construct a HeapSnapshotLoader. We can
simplify the code by introducing a 'createLoader' disposition instead.

It's also unclear why the HeapsnapshotWorkerDispatcher is not allowed
to know about the workers architecture, as the loader is passed in
as a property path relative to `globalThis`. The class must be
available in the worker though for this to work, so it's unclear why
this indirection is needed.

R=alexrudenko@chromium.org

Bug: 1442410
Change-Id: I5649d4c57ae2aeff2b8ae87c03763902a9e97f27
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4853597
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>