[legacy] Use the '-entrypoint.js' directly for the heap snapshot worker
Currently we use the '-legacy.js' file as the entrypoint which also
loads '-entrypoint.js'.

The '-legacy.js' file is only needed for web tests though and we can
get around that by putting the HeapSnapshotWorker and HeapsnapshotModel
on `globalThis` when 'evaluateForTest' is called.

This CL also updates the global usage of HeapSnapshotWorker in the
test harness with an import.

Once this CL has landed, we can update and re-enable the corresponding
web test.

R=alexrudenko@chromium.org

Bug: 1442410
Change-Id: Ida48e461b31a9eab22af2c19c1847f36d87c594e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4853520
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Auto-Submit: Simon Zünd <szuend@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>