Use root target when tracing
This CL migrates DevTools to use the root target when making a recording
via the Tracing domain.

As a by-product of this, when digging into the code with andoli@, we
found a lot of dead code in TimelineController.ts. This code dealt with
the scenario where we are gathering profiles from the Profiler domain.
However, in this situation now the TimelineController is not even
created; we deal with CPU Profile recordings in TimelinePanel.ts.

Therefore, we are able to remove anything from TimelineController
relating to CPU Profiles, hence the code removal in this CL.

Bug: 1478973
Change-Id: I6186ef278ea00eb9967fecad76f04a78689cfa5d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4797743
Commit-Queue: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>