Create a SourceMapsResolver for the new trace engine data
This CL creates (but does not use) a SourceMapsResolver class to manage
resolving node function names from source maps, using trace data from
the new engine.

The code is very similar to the code that exits in PerformanceModel and
the approach we take has not changed.

Bug: 1475936
Change-Id: Ic662e9e15fcca2404bd05901b06d7f46e9acfcfb
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4866347
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>