Fix broken perf panel when user scrolls with no mouse move
If the user scrolls the panel before moving their mouse at all
(unlikely, but I have managed it a few times!), then we would end up
setting the bounds of the window to [NaN, NaN], because in ChartViewport
the mouseX position would be undefined.

This CL fixes this by changing the type definition to acknowledge that
it might be undefined, and if it is, we just use position 0 to zoom
from. This avoids the panel breaking and initiates a zoom.

Fixed: 1484491
Change-Id: I2733d5ac72f8d898322b8258a6616c0ed1ec0e94
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4876634
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>