[coverage] Add unit tests for CoverageDecorationManager
This CL adds some basic unit tests for CoverageDecorationManager in
preparation for fixing coverage for in-place formatted files.

The CL also contains a small refactoring: Instead of using the
global singletons we pass them by constructor now.
Changing this broke the "CoverageView_test.ts" which relied on
other tests booting up the bindings, so we do that explicitly
in the test itself now.

R=jarin@chromium.org

Bug: 1464974
Change-Id: I0c4c7c8eb985f44e0689adc489d2a2934c7744c5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4892800
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>